chore: merge websocket files
